Create lower level -types libraries to reuse feature structs without circular deps
