+++ .github/CODEOWNERS
@@ -52,12 +52,14 @@ crates/bitwarden-policies/** @bitwarden/team-admin-console-dev
 crates/bitwarden-random/** @bitwarden/team-key-management-dev
 crates/bitwarden-sensitive-value/** @bitwarden/team-key-management-dev
 crates/bitwarden-send/** @bitwarden/team-tools-dev
+crates/bitwarden-send-types/** @bitwarden/team-tools-dev
 crates/bitwarden-shared-unlock/** @bitwarden/team-key-management-dev
 crates/bitwarden-ssh/** @bitwarden/team-tools-dev
 crates/bitwarden-state-bridge-macro/** @bitwarden/team-key-management-dev
 crates/bitwarden-unlock/** @bitwarden/team-key-management-dev
 crates/bitwarden-user-crypto-management/** @bitwarden/team-key-management-dev
 crates/bitwarden-vault/** @bitwarden/team-vault-dev
+crates/bitwarden-vault-types/** @bitwarden/team-vault-dev
 
 # Team-owned folders in other crates (to be avoided if possible)
 crates/bitwarden-wasm-internal/integration-tests/tests/organizations/** @bitwarden/team-admin-console-dev
