+++ crates/bitwarden-send-types/src/lib.rs
@@ -1,0 +1,7 @@
+#![doc = include_str!("../README.md")]
+
+#[cfg(feature = "uniffi")]
+uniffi::setup_scaffolding!();
+
+mod send_type;
+pub use send_type::SendType;
