+++ crates/bitwarden-send/src/send.rs
@@ -10,6 +10,7 @@ use bitwarden_crypto::{
     OctetStreamBytes, PrimitiveEncryptable, generate_random_bytes,
 };
 use bitwarden_encoding::{B64, B64Url};
+use bitwarden_send_types::SendType;
 use bitwarden_uuid::uuid_newtype;
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
@@ -81,20 +82,6 @@ pub struct SendTextView {
     pub text: Option<String>,
     /// Whether the text is hidden-by-default (masked as ********).
     pub hidden: bool,
-}
-
-/// The type of Send, either text or file
-#[derive(Clone, Copy, Serialize_repr, Deserialize_repr, Debug, PartialEq)]
-#[repr(u8)]
-#[cfg_attr(feature = "uniffi", derive(uniffi::Enum))]
-#[cfg_attr(feature = "wasm", wasm_bindgen)]
-pub enum SendType {
-    /// Text-based send
-    Text = 0,
-    /// File-based send
-    File = 1,
-    /// Item-based send
-    Item = 2,
 }
 
 /// Indicates the authentication strategy to use when accessing a Send
@@ -647,31 +634,6 @@ impl TryFrom<SendResponseModel> for Send {
             emails: send.emails,
             auth_type,
         })
-    }
-}
-
-impl TryFrom<bitwarden_api_api::models::SendType> for SendType {
-    type Error = bitwarden_core::MissingFieldError;
-
-    fn try_from(t: bitwarden_api_api::models::SendType) -> Result<Self, Self::Error> {
-        Ok(match t {
-            bitwarden_api_api::models::SendType::Text => SendType::Text,
-            bitwarden_api_api::models::SendType::File => SendType::File,
-            bitwarden_api_api::models::SendType::Item => SendType::Item,
-            bitwarden_api_api::models::SendType::__Unknown(_) => {
-                return Err(bitwarden_core::MissingFieldError("type"));
-            }
-        })
-    }
-}
-
-impl From<SendType> for bitwarden_api_api::models::SendType {
-    fn from(t: SendType) -> Self {
-        match t {
-            SendType::Text => bitwarden_api_api::models::SendType::Text,
-            SendType::File => bitwarden_api_api::models::SendType::File,
-            SendType::Item => bitwarden_api_api::models::SendType::Item,
-        }
     }
 }
