+++ crates/bitwarden-vault-types/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "bitwarden-vault-types"
+description = """
+Shared data types for the Bitwarden vault.
+Internal crate for the bitwarden crate. Do not use.
+"""
+
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+readme.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license = "GPL-3.0-only OR LicenseRef-Bitwarden-SDK"
+keywords.workspace = true
+
+[features]
+uniffi = ["dep:uniffi"] # Uniffi bindings
+wasm = ["dep:wasm-bindgen"] # WASM support
+
+[dependencies]
+bitwarden-api-api = { workspace = true }
+bitwarden-core = { workspace = true }
+serde = { workspace = true }
+serde_repr = { workspace = true }
+uniffi = { workspace = true, optional = true }
+wasm-bindgen = { workspace = true, optional = true }
+
+[lints]
+workspace = true
