+++ crates/bitwarden-vault-types/src/lib.rs
@@ -1,0 +1,7 @@
+#![doc = include_str!("../README.md")]
+
+#[cfg(feature = "uniffi")]
+uniffi::setup_scaffolding!();
+
+mod uri_match_type;
+pub use uri_match_type::UriMatchType;
