+++ crates/bitwarden-vault/Cargo.toml
@@ -19,6 +19,7 @@ uniffi = [
     "bitwarden-collections/uniffi",
     "bitwarden-core/uniffi",
     "bitwarden-crypto/uniffi",
+    "bitwarden-vault-types/uniffi",
     "dep:uniffi",
 ] # Uniffi bindings
 wasm = [
@@ -26,6 +27,7 @@ wasm = [
     "bitwarden-core/wasm",
     "bitwarden-encoding/wasm",
     "bitwarden-state/wasm",
+    "bitwarden-vault-types/wasm",
     "dep:tsify",
     "dep:wasm-bindgen",
     "dep:wasm-bindgen-futures",
@@ -45,6 +47,7 @@ bitwarden-ssh = { workspace = true }
 bitwarden-state = { workspace = true }
 bitwarden-sync = { workspace = true }
 bitwarden-uuid = { workspace = true }
+bitwarden-vault-types = { workspace = true }
 chrono = { workspace = true }
 data-encoding = { workspace = true }
 futures = { workspace = true }
